requirements.txt
ultralytics
opencv-python
google-generativeai
openai
google-genai
python-dotenv